Add Docker support and restructure application code
