+++ docker-compose.yml
@@ -1,0 +1,19 @@
+version: '3.8'
+
+services:
+  app:
+    build:
+      context: .
+      dockerfile: Dockerfile
+    ports:
+      - "3000:3000"
+    volumes:
+      - ./src:/app/src
+    environment:
+      NODE_ENV: production
+    networks:
+      - backend-network
+
+networks:
+  backend-network:
+    driver: bridge
