+++ src/index.js
@@ -1,0 +1,22 @@
+const express = require('express');
+const app = express();
+const port = 3000;
+
+// Middleware to parse JSON bodies
+app.use(express.json());
+
+// Import routes
+const routes = require('./routes');
+
+// Use routes
+app.use('/', routes);
+
+// Start the server
+app.listen(port, () => {
+  console.log(`Server is running on http://localhost:${port}`);
+});
+
+// Simple ping request
+app.get('/ping', (req, res) => {
+  res.send('pong');
+});
